# pkg/apiserver/docs/swagger.yaml
definitions:
...
  AgentConditionStatus:
    enum:
    - "True"
    - "False"
    - Unknown
    type: string
    x-enum-varnames:
    - ConditionStatusTrue
    - ConditionStatusFalse
    - ConditionStatusUnknown
...
  AgentPackage:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/AgentPackageMetadata'
      spec:
        $ref: '#/definitions/AgentPackageSpec'
      status:
        $ref: '#/definitions/AgentPackageStatus'
    type: object
  AgentPackageMetadata:
    properties:
      attributes:
        $ref: '#/definitions/github_com_minuk-dev_opampcommander_api_v1.Attributes'
      createdAt:
        type: string
      deletedAt:
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
  AgentPackageSpec:
    properties:
      contentHash:
        items:
          type: integer
        type: array
      downloadUrl:
        type: string
      hash:
        items:
          type: integer
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      packageType:
        type: string
      signature:
        items:
          type: integer
        type: array
      version:
        type: string
    type: object
  AgentPackageStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/Condition'
        type: array
    type: object
...
  Condition:
    properties:
      lastTransitionTime:
        type: string
      message:
        type: string
      reason:
        type: string
      status:
        $ref: '#/definitions/AgentConditionStatus'
      type:
        $ref: '#/definitions/ConditionType'
    type: object
  ConditionType:
    enum:
    - Created
    - Updated
    - Deleted
    - Connected
    - Healthy
    - Configured
    - Registered
    type: string
    x-enum-varnames:
    - ConditionTypeCreated
    - ConditionTypeUpdated
    - ConditionTypeDeleted
    - ConditionTypeConnected
    - ConditionTypeHealthy
    - ConditionTypeConfigured
    - ConditionTypeRegistered
...
  ListMeta:
    properties:
      continue:
        type: string
      remainingItemCount:
        type: integer
    type: object
...
  ListResponse-AgentPackage:
    properties:
      apiVersion:
        type: string
      items:
        items:
          $ref: '#/definitions/AgentPackage'
        type: array
      kind:
        type: string
      metadata:
        $ref: '#/definitions/ListMeta'
    type: object
...
  ListResponse-Namespace:
    properties:
      apiVersion:
        type: string
      items:
        items:
          $ref: '#/definitions/Namespace'
        type: array
      kind:
        type: string
      metadata:
        $ref: '#/definitions/ListMeta'
    type: object
...
  Namespace:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/NamespaceMetadata'
      status:
        $ref: '#/definitions/NamespaceStatus'
    type: object
  NamespaceMetadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      createdAt:
        type: string
      deletedAt:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  NamespaceStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/Condition'
        type: array
    type: object
...
  github_com_minuk-dev_opampcommander_api_v1.Attributes:
    additionalProperties:
      type: string
    type: object
...
info:
  contact: {}
  description: This is the API server for OpAMP Commander, providing endpoints for
    managing OpAMP agents.
  termsOfService: http://swagger.io/terms/
  title: OpAMP Commander API Server
  version: "1.0"
paths:
...
  /api/v1/namespaces:
    get:
      description: Retrieve a list of namespaces.
      parameters:
      - description: Maximum number of namespaces to return
        in: query
        name: limit
        type: integer
      - description: Token to continue listing namespaces
        in: query
        name: continue
        type: string
      - description: Include soft-deleted namespaces
        in: query
        name: includeDeleted
        type: boolean
      - description: Label selector, e.g. env=prod,tier notin (canary,dev)
        in: query
        name: labelSelector
        type: string
      - description: Field selector (this resource exposes no selectable fields)
        in: query
        name: fieldSelector
        type: string
      - description: Case-sensitive name prefix filter
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListResponse-Namespace'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Namespaces
      tags:
      - namespace
...
  /api/v1/namespaces/{namespace}/agentpackages:
    get:
      description: Retrieve a list of agent packages.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Maximum number of agent packages to return
        in: query
        name: limit
        type: integer
      - description: Token to continue listing agent packages
        in: query
        name: continue
        type: string
      - description: Include soft-deleted agent packages
        in: query
        name: includeDeleted
        type: boolean
      - description: Label selector, e.g. env=prod,tier notin (canary,dev)
        in: query
        name: labelSelector
        type: string
      - description: 'Fields: metadata.namespace, spec.packageType, spec.version'
        in: query
        name: fieldSelector
        type: string
      - description: Case-sensitive name prefix filter
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListResponse-AgentPackage'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Agent Packages
      tags:
      - agentpackage
    post:
      consumes:
      - application/json
      description: Create a new agent package.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Agent Package to create
        in: body
        name: agentPackage
        required: true
        schema:
          $ref: '#/definitions/AgentPackage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AgentPackage'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create Agent Package
      tags:
      - agentpackage
  /api/v1/namespaces/{namespace}/agentpackages/{name}:
    delete:
      description: Delete an agent package by its name.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the agent package
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete Agent Package
      tags:
      - agentpackage
    get:
      description: Retrieve an agent package by its name.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the agent package
        in: path
        name: name
        required: true
        type: string
      - description: Include soft-deleted agent package
        in: query
        name: includeDeleted
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AgentPackage'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Agent Package
      tags:
      - agentpackage
    put:
      consumes:
      - application/json
      description: Update an existing agent package.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the agent package
        in: path
        name: name
        required: true
        type: string
      - description: Updated Agent Package
        in: body
        name: agentPackage
        required: true
        schema:
          $ref: '#/definitions/AgentPackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AgentPackage'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update Agent Package
      tags:
      - agentpackage
...
swagger: "2.0"